src/third_party/blink/web_tests/external/wpt/fetch/sec-metadata/redirect/same-origin-redirect.tentative.https.sub-expected.txt

This is a testharness.js-based test.
PASS Same-Origin -> Same-Origin redirect
FAIL Same-Origin -> Same-Site redirect assert_equals: site expected "same-site" but got "same-origin"
FAIL Same-Origin -> Cross-Site redirect assert_equals: site expected "cross-site" but got "same-origin"
Harness: the test ran to completion.
